add unit tests for weather forecast ViewModel and Presenter

+++ app/src/main/java/com/cookiesmile/mnml_weather/data/api/response/forecast/inner/Clouds.java
@@ -13,5 +13,5 @@ public abstract class Clouds {
   }
 
   @Json(name = "all")
-  public abstract long all();
+  public abstract double all();
 }
